+++ Antlr3/Tool/RuleLabelScope.cs
@@ -4,7 +4,7 @@
  * All rights reserved.
  *
  * Conversion to C#:
- * Copyright (c) 2008 Sam Harwell, Pixel Mine, Inc.
+ * Copyright (c) 2008-2009 Sam Harwell, Pixel Mine, Inc.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -102,7 +102,7 @@ namespace Antlr3.Tool
         public override Attribute GetAttribute( string name )
         {
             AttributeScope rulePropertiesScope =
-                RuleLabelScope.grammarTypeToRulePropertiesScope[grammar.type];
+                RuleLabelScope.grammarTypeToRulePropertiesScope[(int)grammar.type];
             if ( rulePropertiesScope.GetAttribute( name ) != null )
             {
                 return rulePropertiesScope.GetAttribute( name );
